more unsigned long where size_t should be used
memory_managers/array_grid.cc:549:6:   required from here
memory_managers/hole_base.h:232:13: error: cast from 'short int*' to
'long unsigned int' loses precision [-fpermissive]
   s.put_hex((unsigned long)data);
             ^~~~~~~~~~~~~~~~~~~

+++ src/memory_managers/hole_base.h
@@ -229,7 +229,7 @@ template <class INT>
 void MEDDLY::hole_manager<INT>::showInternal(output &s) const
 {
   s << "  data pointer: ";
-  s.put_hex((unsigned long)data);
+  s.put_hex((size_t)data);
   s << "\n";
   s << "  data_alloc: " << data_alloc << "\n";
   s << "  last_used_slot: " << last_used_slot << "\n\n";
